+++ demo/test/components/previews/primer/auto_complete_preview.rb
@@ -3,6 +3,7 @@ module Primer
     def default
       render(Primer::AutoComplete.new(id: "test-id", src: "/auto_complete", position: :relative)) do |c|
         c.input(type: :text, name: "input", "aria-label": "Select a fruit")
+        c.results("aria-label": "Select a fruit")
       end
     end
 
@@ -10,6 +11,7 @@ module Primer
       render(Primer::AutoComplete.new(id: "test-id", src: "/auto_complete", position: :relative)) do |c|
         c.icon(icon: :search)
         c.input(type: :text, name: "input", "aria-label": "Select a fruit")
+        c.results("aria-label": "Select a fruit")
       end
     end
   end