+++ demo/test/components/previews/primer/docs/auto_complete_preview.rb
@@ -3,6 +3,8 @@ module Primer
     class AutoCompletePreview < ViewComponent::Preview
       def default; end
 
+      def with_aria_label; end
+
       def with_custom_classes_for_the_results; end
 
       def with_icon; end